requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.2
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-countries==7.2.1
django-heroku==0.3.1
django-jsonfield==1.4.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
jsonfield==3.1.0
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
Pillow==8.3.1
pipenv==2020.11.15
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==1.7.1
pylint==2.7.4
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rest-framework-generic-relations==2.0.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.4
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
